% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_age_groups.R
\name{assign_age_groups}
\alias{assign_age_groups}
\title{Assign age groups}
\usage{
assign_age_groups(df, groups = NULL, aggregation_function = sum)
}
\arguments{
\item{df}{a data table with a single-year-of-age column called \code{age}}

\item{aggregation.function}{"sum" or "mean", use mean for life expectancy}
}
\value{
a data table summarised by the ONS age groups in a column called
\code{age_group}
}
\description{
Since the publicly available alcohol-specific deaths data is only available
by age with age groups we have to group the single-year of age data in the
same way.
}
