% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_poisoning_deaths.R
\name{get_non_poisoning_deaths}
\alias{get_non_poisoning_deaths}
\title{Load NDTMS-ONS mortality linked dataset}
\usage{
get_non_poisoning_deaths()
}
\description{
Load NDTMS-ONS mortality linked dataset
}
