% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_alcohol_deaths.R
\name{merge_alcohol_deaths}
\alias{merge_alcohol_deaths}
\title{Merge alcohol deaths}
\usage{
merge_alcohol_deaths()
}
\arguments{
\item{data}{A data frame produced by \code{get_non_poisoning_deaths}}

\item{years}{Vector of years to include in the analysis.}

\item{groups}{Grouping variable:
"area_name", "age", "sex", "death_cause", or NULL. Always grouped by year.}

\item{by_treatment_status}{Whether to group by treatment status.}

\item{substance_group}{Drugs only ("drugs"), alcohol only ("alcohol"),
or both ("both")}
}
\value{
A data table with grouped and summarised treatment data.
}
\description{
Get ONS alcohol-specific death data by age and merge with additional deaths
related to alcohol use as determined by analysis of the NDMTS-ONS linked
mortality dataset.
}
