% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_non_poisoning_data.R
\name{process_non_poisoning_data}
\alias{process_non_poisoning_data}
\title{Process deaths in treatment data}
\usage{
process_non_poisoning_data(
  data,
  years = c(2021, 2022),
  groups = NULL,
  by_treatment_status = TRUE,
  substance_group = "drugs"
)
}
\arguments{
\item{data}{A data frame produced by \code{get_non_poisoning_deaths}}

\item{years}{Vector of years to include in the analysis.}

\item{groups}{Grouping variable:
"area_name", "age", "sex", "death_cause", or NULL. Always grouped by year}

\item{by_treatment_status}{Whether to group by treatment status.}

\item{substance_group}{Drugs only ("drugs"),
alcohol only ("alcohol"), or both ("both")}
}
\value{
A data table with grouped and summarised treatment data.
}
\description{
Process deaths in treatment data
}
