% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_non_poisoning_deaths.R
\name{summarise_non_poisoning_deaths}
\alias{summarise_non_poisoning_deaths}
\title{Summarise non-poisoning deaths data}
\usage{
summarise_non_poisoning_deaths(data, years = NULL, groups)
}
\arguments{
\item{data}{A data frame, the output of get_non_poisoning_deaths()}

\item{years}{A numeric vector of years to select}

\item{groups}{A string vector:
area, death_cause, age, age_group, sex, treatment_status}
}
\description{
Summarise non-poisoning deaths data
}
